Deduplicate workflow response builders
Remove duplicated workflow response builders for MCP features that only
delegate to the shared rule DistSQL payload builder.

Rename the remaining encrypt and mask response builders with feature
prefixes because they include feature-specific masked property previews.
